Fix bug in save policy line
The save policy line function was saving all parts of the policy line
to v0 in the model. The new function will properly split the rule
parts into different columns.

Tests are added for `save_policy` and `add_policy` functions.

Adds custom `__str__` and `__repr__` functions with tests for reuse.

Streamlines testing using in-memory sqlite instead of a file.